Commit before upgrade to llvm-4.0

+++ lib/epp/EPPPathPrinter.cpp
@@ -61,12 +61,14 @@ bool EPPPathPrinter::runOnModule(Module &M) {
     try {
         string Line;
         while (getline(InFile, Line)) {
-            uint32_t FunctionId = 0, NumberOfPaths = 0;
+            uint64_t FunctionId = 0, NumberOfPaths = 0;
             stringstream SS(Line);
             SS >> FunctionId >> NumberOfPaths;
 
             // If no paths have been executed for this function,
-            // then skip it altogether.
+            // then skip it altogether. A continue over here is fine,
+            // since there are no lines for the paths themselves and 
+            // the next line we expect is for one FunctionId and NumberOfPaths.
 
             if (NumberOfPaths == 0)
                 continue;
@@ -78,6 +80,8 @@ bool EPPPathPrinter::runOnModule(Module &M) {
             vector<Path> Paths;
             for (uint32_t I = 0; I < NumberOfPaths; I++) {
                 getline(InFile, Line);
+
+
                 stringstream SS(Line);
                 string PathIdStr;
                 uint64_t PathExecFreq;
